refactor: rename module options
