+++ packages/core/src/modules/discover-features/protocol/v2/V2DiscoverFeaturesService.ts
@@ -73,7 +73,7 @@ export class V2DiscoverFeaturesService extends DiscoverFeaturesService {
 
     // Process query and send responde automatically if configured to do so, otherwise
     // just send the event and let controller decide
-    if (this.discoverFeaturesModuleConfig.autoAcceptDiscoverFeatureQueries) {
+    if (this.discoverFeaturesModuleConfig.autoAcceptQueries) {
       return await this.createDisclosure({
         threadId,
         disclosureQueries: queries,
